Add DDL and ETL scripts and docs
The docs are collected from the OMOP website. The etl scripts are a
combination of provided scripts and custom ones. The provided scripts
had to be modified to both work with PostgreSQL (they were written for
Oracle) and also to accept the OSIM2 data which is in an older version
of the CDM. There are some known issues recorded in etl/KNOWN_ISSUES.md
and the process is described in etl/README.md