Create Django models for the CDMV4
Models are created from a syncdb command after the DDL and ETL process.
South initial migration is created, which could be used instead of the
DDL routines in the ETL folder. Comments from the OMOP DDL are added as
help_text to fields in the Django models, but no verbose names are
defined.