Start PEDSnet development
This commit breaks from previous development, which was toward a
demo-able app built on the OSIM2 simulated dataset. The new development
path is towards a production-capable app built on the OMOP CDMV4 data
model for the PEDSnet Data Coordinating Center. In that spirit, this
commit removes all migrations and starts over with new models, beginning
with person, provider, care_site, organization, location, visit,
condition, procedure, and drug. These models are intended to sit on top
of materialized views, the code for which is in the PEDSnet DCC repo.

This commit also updates the project dependencies, including cilantro,
and makes changes to the settings to accomodate the django-haystack
upgrade. The base.html template is modified to include the bootstrap css
files as cilantro is no longer packaged with them.

The main.js project bootstrap file is modified to customize the type of
two fields (person_id and provider_id) as strings, in order to convince
cilantro to use pick-list controls for those fields. The avocado
datafields and dataconcepts are configured and stored in a fixture with
paired metadata migration.

Fixes #4

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>