Update DDL code
Update the DDL code to match the progress in the PEDSnet repositories.
This update should still allow for use of the OSIM2 datasets. For that
reason, the remove_CDMV2_fields.sh script is added to pre-process OSIM2
data for use in the CDMV4 tables. The README is also modified to add
this step, but needs more modification and a possible re-working of
startup steps.

Fixes #16
Fixes #18

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>