Fix OSIM data load script, start using OSIM SNOMED
The OSIM2_data_load.psql script had a word order error that caused it to
fail, this is fixed. That script as well as the pre-processing scripts (remove
pipes and quotes and remove CDMV2 fields scripts) have been changed to
target the OSIM2_1M_MSLR_SNOMED_0_CSV dataset because it is much more
manageable than the 10M dataset used previously.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>